backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
SQLAlchemy==2.0.36
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.7.0
openai==1.58.1
opencv-python-headless==4.10.0.84
faster-whisper==1.1.1
pymysql==1.1.1
requests==2.32.3
python-jose[cryptography]==3.3.0
email-validator==2.2.0
